optimized_trading_bot/requirements.txt
# Optimized Trading Bot Dependencies
# Production-ready, version-locked dependencies for stability and security

# Core Trading & Market Data
alpaca-py                         # Alpaca API client
pandas==2.1.4                     # Data manipulation and analysis
numpy==1.24.4                     # Numerical computing
ta==0.10.2                        # Technical analysis indicators
yfinance==0.2.36                  # Yahoo Finance data (backup)

# Async & Performance
asyncio==3.4.3                    # Async programming support
aiohttp==3.9.3                    # Async HTTP client/server
aiofiles==23.2.1                  # Async file operations
uvloop==0.19.0                    # Fast async event loop (Unix only)

# Data Storage & Caching
redis==5.0.1                      # In-memory data store for caching
sqlalchemy==2.0.25                # SQL toolkit and ORM
alembic==1.13.1                   # Database migration tool
aiosqlite==0.19.0                 # Async SQLite interface

# Configuration & Validation
pydantic==2.5.3                   # Data validation and settings
pydantic-settings==2.1.0          # Settings management
python-dotenv==1.0.0              # Environment variable management
jsonschema==4.20.0                # JSON schema validation

# Machine Learning (Optional)
scikit-learn==1.3.2               # Machine learning library
numpy-financial==1.0.0            # Financial calculations
# ta-lib                             # Technical analysis library (requires compilation)
# Note: ta-lib is optional and may cause Docker build issues
# The bot can run without it using the 'ta' library instead

# Monitoring & Observability
prometheus-client==0.19.0         # Prometheus metrics
psutil==5.9.6                     # System monitoring
structlog==23.2.0                 # Structured logging
rich==13.7.0                      # Rich console output

# Web Dashboard & API
fastapi==0.108.0                  # Modern web framework
flask==3.0.0                      # Lightweight web framework
flask-cors==4.0.0                 # CORS support for Flask
uvicorn==0.25.0                   # ASGI server
websockets                       # WebSocket support
jinja2==3.1.2                     # Template engine
plotly==5.18.0                    # Interactive plots

# Notifications
twilio==8.11.1                    # SMS notifications
python-telegram-bot==20.7         # Telegram bot API
requests==2.31.0                  # HTTP requests

# Security & Encryption
cryptography>=41,<46             # Cryptographic recipes
bcrypt==4.1.2                     # Password hashing
python-jose[cryptography]==3.3.0  # JWT handling

# Testing & Development
pytest==7.4.4                     # Testing framework
pytest-asyncio==0.23.2            # Async testing support
pytest-cov==4.1.0                 # Coverage reporting
pytest-mock==3.12.0               # Mocking support
black==23.12.1                    # Code formatting
flake8==7.0.0                     # Code linting
mypy==1.8.0                       # Static type checking
isort==5.13.2                     # Import sorting

# Utilities
click==8.1.7                      # Command line interface
python-dateutil==2.8.2            # Date utilities
pytz==2023.3                      # Timezone support
schedule==1.2.0                   # Job scheduling
tenacity==8.2.3                   # Retry library

# Cloud & Deployment (Optional)
docker==7.0.0                     # Docker API client
kubernetes==28.1.0                # Kubernetes API client
boto3==1.34.0                     # AWS SDK
google-cloud-storage==2.10.0      # Google Cloud Storage

# Monitoring & Tunneling
pyngrok==7.0.0                    # ngrok tunnel management
websocket-client==1.7.0           # WebSocket client

# Data Analysis & Visualization
matplotlib==3.8.2                 # Plotting library
seaborn==0.13.0                   # Statistical data visualization
jupyter==1.0.0                    # Jupyter notebook (development)

# Financial Data Sources
alpha-vantage==2.3.1              # Alpha Vantage API
fredapi==0.5.1                    # Federal Reserve Economic Data
quandl==3.7.0                     # Quandl financial data

# Performance & Profiling
line-profiler==4.1.1              # Line-by-line profiling
memory-profiler==0.61.0           # Memory usage profiling
py-spy==0.3.14                    # Sampling profiler

# Database Connectors
psycopg2-binary==2.9.9            # PostgreSQL adapter
pymongo==4.6.1                    # MongoDB driver

# Environment-specific extras
# Install with: pip install -r requirements.txt -r requirements-dev.txt (for development)
# Production installs should exclude development dependencies